use shouldcomponentupdate to optimize explorer
