+++ .eslintrc.json
@@ -18,7 +18,7 @@
   },
   "extends": "airbnb",
   "rules": {
-    "max-len": ["error", { "code": 150, "ignoreTemplateLiterals": true, "ignoreStrings": true }],
+    "max-len": ["error", { "code": 180, "ignoreTemplateLiterals": true, "ignoreStrings": true }],
     "no-shadow": 0,
     "no-mixed-operators": ["error", {
       "allowSamePrecedence": true
@@ -34,8 +34,8 @@
     "jsx-a11y/no-static-element-interactions": [0],
     "jsx-a11y/img-has-alt": 1,
     "react/no-array-index-key": 1,
-    "react/prop-types": 1,
-    "react/require-default-props": 1,
+    "react/prop-types": 0,
+    "react/require-default-props": 0,
     "react/jsx-key": 2,
     "react/forbid-prop-types": [2, { "forbid": ["any"] }]
   }
