+++ src/components/Explorer/Explorer.jsx
@@ -8,43 +8,28 @@ import fetch from 'isomorphic-fetch';
 import Spinner from 'components/Spinner';
 import RaisedButton from 'material-ui/RaisedButton';
 import Toggle from 'material-ui/Toggle';
-import {
-  Link,
-}
-from 'react-router';
-import Helmet from 'react-helmet';
 import strings from 'lang';
+import Helmet from 'react-helmet';
 import {
   getScript,
-  transformations,
-  formatSeconds,
 }
 from 'utility';
-import Table from 'components/Table';
 import Heading from 'components/Heading';
-import itemData from 'dotaconstants/build/items.json';
 import {
   getProPlayers,
   getLeagues,
   getTeams,
 }
 from 'actions';
-import {
-  TablePercent,
-  inflictorWithValue,
-}
-from 'components/Visualizations';
-import { IconRadiant, IconDire } from 'components/Icons';
-import matchStyles from 'components/Match/Match.css';
+
 import querystring from 'querystring';
 import json2csv from 'json2csv';
-import heroes from 'dotaconstants/build/heroes.json';
 import queryTemplate from './queryTemplate';
 import ExplorerFormField from './ExplorerFormField';
 import ExplorerOutputButton from './ExplorerOutputButton';
+import ExplorerOutputSection from './ExplorerOutputSection';
 import fields from './fields';
 import autocomplete from './autocomplete';
-import redrawGraphs from './redrawGraphs';
 import styles from './Explorer.css';
 
 // TODO split picks/bans by phase
@@ -62,8 +47,6 @@ import styles from './Explorer.css';
 
 const playerMapping = {};
 const teamMapping = {};
-let currRows = null;
-let currFormat = null;
 
 function jsonResponse(response) {
   return response.json();
@@ -79,137 +62,47 @@ function expandBuilderState(builder, fields) {
   return expandedBuilder;
 }
 
-function resolveId(key, value, mappings) {
-  if (key === 'hero_id') {
-    return (heroes[value] || {}).localized_name;
-  } else if (key === 'account_id') {
-    return mappings.playerMapping[value];
-  } else if (key === 'team_id') {
-    return mappings.teamMapping[value];
-  }
-  return value;
-}
-
-const ExplorerControlSection = ({ showEditor, toggleEditor, expandedFields, handleFieldUpdate, builder }) => {
-    return (<div>
-      <div style={{ width: '180px', margin: '10px' }}>
-        <div>{/* drawOmnibox(this, expandedFields)*/}</div>
-        <Toggle
-          label={strings.explorer_toggle_sql}
-          defaultToggled={showEditor}
-          onToggle={toggleEditor}
-        />
-      </div>
-      <div style={{ display: showEditor ? 'none' : 'flex' }} className={styles.formGroup}>
-        <ExplorerFormField label={strings.explorer_select} fields={expandedFields} builderField="select" handleFieldUpdate={handleFieldUpdate} builder={builder} />
-        <ExplorerFormField label={strings.explorer_group_by} fields={expandedFields} builderField="group" handleFieldUpdate={handleFieldUpdate} builder={builder} />
-        <ExplorerFormField label={strings.explorer_hero} fields={expandedFields} builderField="hero" handleFieldUpdate={handleFieldUpdate} builder={builder} />
-        <ExplorerFormField label={strings.explorer_player} fields={expandedFields} builderField="player" handleFieldUpdate={handleFieldUpdate} builder={builder} />
-        <ExplorerFormField label={strings.explorer_team} fields={expandedFields} builderField="team" handleFieldUpdate={handleFieldUpdate} builder={builder} />
-        <ExplorerFormField label={strings.explorer_league} fields={expandedFields} builderField="league" handleFieldUpdate={handleFieldUpdate} builder={builder} />
-        <ExplorerFormField label={strings.explorer_region} fields={expandedFields} builderField="region" handleFieldUpdate={handleFieldUpdate} builder={builder} />
-        <ExplorerFormField label={strings.explorer_patch} fields={expandedFields} builderField="patch" handleFieldUpdate={handleFieldUpdate} builder={builder} />
-        <ExplorerFormField label={strings.explorer_duration} fields={expandedFields} builderField="duration" handleFieldUpdate={handleFieldUpdate} builder={builder} />
-        <ExplorerFormField label={strings.explorer_side} fields={expandedFields} builderField="side" handleFieldUpdate={handleFieldUpdate} builder={builder} />
-        <ExplorerFormField label={strings.th_result} fields={expandedFields} builderField="result" handleFieldUpdate={handleFieldUpdate} builder={builder} />
-        <ExplorerFormField
-          label={strings.explorer_player_purchased}
-          fields={expandedFields}
-          builderField="playerPurchased"
-          handleFieldUpdate={handleFieldUpdate} builder={builder}
-        />
-        {/* <ExplorerFormField label={strings.explorer_lane_pos} fields={expandedFields} builderField="lanePos" handleFieldUpdate={handleFieldUpdate} builder={builder} />*/}
-        <ExplorerFormField label={strings.explorer_min_date} builderField="minDate" handleFieldUpdate={handleFieldUpdate} builder={builder} isDateField />
-        <ExplorerFormField label={strings.explorer_max_date} builderField="maxDate" handleFieldUpdate={handleFieldUpdate} builder={builder} isDateField />
-      </div>
-      <div style={{ display: showEditor ? 'block' : 'none' }}>
-        <div
-          id={'editor'}
-          style={{
-            height: 100,
-            width: '100%',
-          }}
-        />
-      </div>
-    </div>);
-  };
-
-const ExplorerOutputSection = ({ rows, fields, expandedBuilder, teamMapping, playerMapping, format }) => {
-  console.log('redrawoutput');
-    setTimeout(() => {
-      if ((currRows !== rows || currFormat !== format) && fields) {
-        const firstCol = fields[0].name;
-        redrawGraphs(rows.map(row => ({
-          ...row,
-          [firstCol]: resolveId(firstCol, row[firstCol], { teamMapping, playerMapping }) }
-      )), firstCol, (expandedBuilder.select && expandedBuilder.select.key) || strings.th_count);
-        currRows = rows;
-        currFormat = format;
-      }
-    }, 100);
-    if (format === 'donut') {
-      return <div id="donut" />;
-    } else if (format === 'bar') {
-      return <div id="bar" />;
-    } else if (format === 'timeseries') {
-      return <div id="timeseries" />;
-    }
-    return (
-      <Table
-        data={(rows || []).slice(0, 1000)}
-        columns={(fields || []).map(column => ({
-          displayName: column.name,
-          field: column.name,
-        })).map(column => ({
-          ...column,
-          displayFn: (row, col, field) => {
-            if (column.field === 'match_id') {
-              return <Link to={`/matches/${field}`}>{field}</Link>;
-            } else if (column.field.indexOf('hero_id') === 0) {
-              return transformations.hero_id(row, col, field);
-            } else if (column.field.indexOf('account_id') === 0) {
-              return <Link to={`/players/${field}`}>{playerMapping[field] || field}</Link>;
-            } else if (column.field === 'winrate') {
-              return (field >= 0 && field <= 1 ? <TablePercent
-              percent={Number((field * 100).toFixed(2))}
-            /> : null);
-            } else if (column.field === 'adj_winrate') {
-          /*
-          const phat = field;
-          const z = 1.96;
-          const n = row.count;
-          return ((phat + z * z / (2 * n) - z * Math.sqrt((phat * (1 - phat) + z * z / (4 * n)) / n)) / (1 + z * z / n)).toFixed(2);
-          */
-            return field;
-          } else if (column.field === 'rune_id') {
-            return strings[`rune_${field}`];
-          } else if (column.field === 'item_name') {
-            return itemData[field] ? itemData[field].dname : field;
-          } else if (column.field === 'team_id') {
-            return teamMapping[field] || field;
-          } else if (column.field === 'time' || (column.field === 'avg' && expandedBuilder.select && expandedBuilder.select.formatSeconds)) {
-            return formatSeconds(field);
-          } else if (column.field === 'inflictor') {
-            return <span>{inflictorWithValue(field)} {field}</span>;
-          } else if (column.field === 'win') {
-            return <span className={field ? styles.textSuccess : styles.textDanger}>{field ? strings.td_win : strings.td_loss}</span>;
-          } else if (column.field === 'is_radiant') {
-            return field
-            ? <span className={matchStyles.teamIconContainer}><IconRadiant className={matchStyles.iconRadiant} />{strings.general_radiant}</span>
-            : <span className={matchStyles.teamIconContainer}><IconDire className={matchStyles.iconDire} />{strings.general_dire}</span>;
-          } else if (column.field === 'start_time') {
-            return (new Date(field * 1000)).toLocaleDateString('en-US', {
-              day: 'numeric',
-              month: 'short',
-              year: 'numeric',
-            });
-          }
-            return typeof field === 'string' ? field : JSON.stringify(field);
-          },
-          sortFn: row => (isNaN(Number(row[column.field])) ? row[column.field] : Number(row[column.field])),
-        }))}
-      />);
-  };
+const ExplorerControlSection = ({ showEditor, toggleEditor, expandedFields, handleFieldUpdate, builder }) => (<div>
+  <div style={{ width: '180px', margin: '10px' }}>
+    <div>{/* drawOmnibox(this, expandedFields)*/}</div>
+    <Toggle
+      label={strings.explorer_toggle_sql}
+      defaultToggled={showEditor}
+      onToggle={toggleEditor}
+    />
+  </div>
+  <div style={{ display: showEditor ? 'none' : 'flex' }} className={styles.formGroup}>
+    <ExplorerFormField label={strings.explorer_select} fields={expandedFields} builderField="select" handleFieldUpdate={handleFieldUpdate} builder={builder} />
+    <ExplorerFormField label={strings.explorer_group_by} fields={expandedFields} builderField="group" handleFieldUpdate={handleFieldUpdate} builder={builder} />
+    <ExplorerFormField label={strings.explorer_hero} fields={expandedFields} builderField="hero" handleFieldUpdate={handleFieldUpdate} builder={builder} />
+    <ExplorerFormField label={strings.explorer_player} fields={expandedFields} builderField="player" handleFieldUpdate={handleFieldUpdate} builder={builder} />
+    <ExplorerFormField label={strings.explorer_team} fields={expandedFields} builderField="team" handleFieldUpdate={handleFieldUpdate} builder={builder} />
+    <ExplorerFormField label={strings.explorer_league} fields={expandedFields} builderField="league" handleFieldUpdate={handleFieldUpdate} builder={builder} />
+    <ExplorerFormField label={strings.explorer_region} fields={expandedFields} builderField="region" handleFieldUpdate={handleFieldUpdate} builder={builder} />
+    <ExplorerFormField label={strings.explorer_patch} fields={expandedFields} builderField="patch" handleFieldUpdate={handleFieldUpdate} builder={builder} />
+    <ExplorerFormField label={strings.explorer_duration} fields={expandedFields} builderField="duration" handleFieldUpdate={handleFieldUpdate} builder={builder} />
+    <ExplorerFormField label={strings.explorer_side} fields={expandedFields} builderField="side" handleFieldUpdate={handleFieldUpdate} builder={builder} />
+    <ExplorerFormField label={strings.th_result} fields={expandedFields} builderField="result" handleFieldUpdate={handleFieldUpdate} builder={builder} />
+    <ExplorerFormField
+      label={strings.explorer_player_purchased}
+      fields={expandedFields}
+      builderField="playerPurchased"
+      handleFieldUpdate={handleFieldUpdate} builder={builder}
+    />
+    {/* <ExplorerFormField label={strings.explorer_lane_pos} fields={expandedFields} builderField="lanePos" handleFieldUpdate={handleFieldUpdate} builder={builder} />*/}
+    <ExplorerFormField label={strings.explorer_min_date} builderField="minDate" handleFieldUpdate={handleFieldUpdate} builder={builder} isDateField />
+    <ExplorerFormField label={strings.explorer_max_date} builderField="maxDate" handleFieldUpdate={handleFieldUpdate} builder={builder} isDateField />
+  </div>
+  <div style={{ display: showEditor ? 'block' : 'none' }}>
+    <div
+      id={'editor'}
+      style={{
+        height: 100,
+        width: '100%',
+      }}
+    />
+  </div>
+</div>);
 
 class Explorer extends React.Component {
   constructor() {
